# dictionary-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary API",
        "description": "Documentação da API de Dicionário",
        "contact": {
            "email": "support@coodesh.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/user/me/history": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Listar histórico de palavras do usuário",
                "description": "Retorna o histórico de palavras pesquisadas pelo usuário com suporte a paginação.",
                "operationId": "987f9560f648cae192c8bfee9dcb4da0",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para a paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico de palavras do usuário retornado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "word": {
                                                        "type": "string",
                                                        "example": "example"
                                                    },
                                                    "added": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-23T12:34:56+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "totalDocs": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totalPages": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "hasNext": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "hasPrev": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de solicitação, parâmetros inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Parâmetros inválidos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Perfil do usuário",
                "description": "Retorna as informações do usuário autenticado.",
                "operationId": "8120197dbf6b5752a156b9868aa8bd4e",
                "responses": {
                    "200": {
                        "description": "Informações do usuário retornadas com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "João da Silva"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "joao@exemplo.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-23T12:34:56+00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-12-23T12:34:56+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário não autenticado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}